Refactored Tab and Tabs into .vue files. Applied basic styling to Tabs.
