+++ src/components/tabs/tab.vue
@@ -1,0 +1,16 @@
+<template>
+
+<div>
+  <slot name="header"></slot>
+  <slot name="body"></slot>
+</div>
+
+</template>
+
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator';
+
+@Component
+export default class Tab extends Vue {}
+
+</script>
